splitted up tests
